Move oraxen.mechanic.magicstaff to oraxen.mechanic.artifact
