+++ src/main/java/jp/mincra/mincramagics/oraxen/mechanic/artifact/MagicStaffMechanicManager.java
@@ -1,4 +1,4 @@
-package jp.mincra.mincramagics.oraxen.mechanic.magicstaff;
+package jp.mincra.mincramagics.oraxen.mechanic.artifact;
 
 import io.th0rgal.oraxen.api.OraxenItems;
 import jp.mincra.mincramagics.MincraMagics;
